+++ r-package/src/R_wrapper.c
@@ -61,12 +61,14 @@ SEXP DownloadURLListR_wrap(SEXP path) {
     return result;
 }
 
-SEXP ConvertR_wrap(SEXP inputDir, SEXP selectedFormats, SEXP tikaAddress) {
+SEXP ConvertR_wrap(SEXP inputDir, SEXP selectedFormats, SEXP tikaAddress, SEXP singleFile, SEXP ocrOnly) {
     const char *c_inputDir = (const char*)inputDir;
     const char *c_selectedFormats = (const char*)selectedFormats;
     const char *c_tikaAddress = (const char*)tikaAddress;
+    const char *c_singleFile = (const char*)singleFile;
+    const char *c_ocrOnly = (const char*)ocrOnly;
 
-    const char *c_result = ConvertR((char *)c_inputDir, (char *)c_selectedFormats, (char *)c_tikaAddress);
+    const char *c_result = ConvertR((char *)c_inputDir, (char *)c_selectedFormats, (char *)c_tikaAddress, (char *)c_singleFile, (char *)c_ocrOnly);
 
     SEXP result = Rf_mkString(c_result);
     PROTECT(result);
@@ -109,7 +111,7 @@ SEXP DownloadURLListR_wrap(SEXP path) {
     return result;
 }
 
-SEXP ConvertR_wrap(SEXP inputDir, SEXP selectedFormats, SEXP tikaAddress) {
+SEXP ConvertR_wrap(SEXP inputDir, SEXP selectedFormats, SEXP tikaAddress, SEXP singleFile, SEXP ocrOnly) {
     SEXP result = Rf_mkString(UNSUPPORTED_PLATFORM_MSG);
     PROTECT(result);
     UNPROTECT(1);